pyproject.toml
[tool.poetry]
name = "trialbot"
version = "0.1.0"
description = ""
authors = ["ginop1 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
gtts = "^2.2.2"
python-dotenv = "^0.17.0"
wikipedia-for-humans = "^0.2.3"
PyNaCl = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
